+++ src/lumberjackth/__init__.py
@@ -8,7 +8,7 @@ from lumberjackth.exceptions import (
     TreeherderNotFoundError,
 )
 
-__version__ = "1.3.0"
+__version__ = "1.4.0"
 __all__ = [
     "LumberjackError",
     "TreeherderAPIError",